refactor(anthropic-vertex): share catalog runtime
