+++ extensions/anthropic-vertex/index.ts
@@ -8,12 +8,8 @@ import {
   buildProviderReplayFamilyHooks,
   resolveClaudeThinkingProfile,
 } from "openclaw/plugin-sdk/provider-model-shared";
-import {
-  hasAnthropicVertexAvailableAuth,
-  mergeImplicitAnthropicVertexProvider,
-  resolveAnthropicVertexConfigApiKey,
-  resolveImplicitAnthropicVertexProvider,
-} from "./api.js";
+import { hasAnthropicVertexAvailableAuth, resolveAnthropicVertexConfigApiKey } from "./api.js";
+import { runAnthropicVertexCatalog } from "./provider-catalog-runtime.js";
 import { normalizeAnthropicVertexResolvedModel } from "./provider-catalog.js";
 
 const PROVIDER_ID = "anthropic-vertex";
@@ -32,20 +28,7 @@ export default definePluginEntry({
       auth: [],
       catalog: {
         order: "simple",
-        run: async (ctx) => {
-          const implicit = resolveImplicitAnthropicVertexProvider({
-            env: ctx.env,
-          });
-          if (!implicit) {
-            return null;
-          }
-          return {
-            provider: mergeImplicitAnthropicVertexProvider({
-              existing: ctx.config.models?.providers?.[PROVIDER_ID],
-              implicit,
-            }),
-          };
-        },
+        run: runAnthropicVertexCatalog,
       },
       resolveConfigApiKey: ({ env }) => resolveAnthropicVertexConfigApiKey(env),
       ...buildProviderReplayFamilyHooks({ family: "native-anthropic-by-model" }),
