+++ extensions/anthropic-vertex/provider-discovery.ts
@@ -3,8 +3,7 @@
  * catalog surfaces that need the provider contract without full plugin entry setup.
  */
 import type { ProviderCatalogContext } from "openclaw/plugin-sdk/provider-catalog-shared";
-import type { ModelProviderConfig } from "openclaw/plugin-sdk/provider-model-shared";
-import { buildAnthropicVertexProvider } from "./provider-catalog.js";
+import { runAnthropicVertexCatalog } from "./provider-catalog-runtime.js";
 import { hasAnthropicVertexAvailableAuth, resolveAnthropicVertexConfigApiKey } from "./region.js";
 
 const PROVIDER_ID = "anthropic-vertex";
@@ -28,48 +27,6 @@ type AnthropicVertexProviderPlugin = {
       }
     | undefined;
 };
-
-function mergeImplicitAnthropicVertexProvider(params: {
-  existing?: ModelProviderConfig;
-  implicit: ModelProviderConfig;
-}) {
-  const { existing, implicit } = params;
-  if (!existing) {
-    return implicit;
-  }
-  return {
-    ...implicit,
-    ...existing,
-    models:
-      Array.isArray(existing.models) && existing.models.length > 0
-        ? existing.models
-        : implicit.models,
-  };
-}
-
-function resolveImplicitAnthropicVertexProvider(params?: { env?: NodeJS.ProcessEnv }) {
-  const env = params?.env ?? process.env;
-  if (!hasAnthropicVertexAvailableAuth(env)) {
-    return null;
-  }
-
-  return buildAnthropicVertexProvider({ env });
-}
-
-async function runAnthropicVertexCatalog(ctx: ProviderCatalogContext) {
-  const implicit = resolveImplicitAnthropicVertexProvider({
-    env: ctx.env,
-  });
-  if (!implicit) {
-    return null;
-  }
-  return {
-    provider: mergeImplicitAnthropicVertexProvider({
-      existing: ctx.config.models?.providers?.[PROVIDER_ID],
-      implicit,
-    }),
-  };
-}
 
 /** Anthropic Vertex provider discovery descriptor. */
 export const anthropicVertexProviderDiscovery: AnthropicVertexProviderPlugin = {
